add 'Desynchronized state' reason
In RFC 9293 3.5.1. Half-Open Connections and Other Anomalies, it says:
"Such connections will automatically become reset if an attempt is made
to send data in either direction."

Then we may notice it also says after the above sentence:
"If at site A the connection no longer exists, then an attempt by the
user at site B to send any data on it will result in the site B TCP
endpoint receiving a reset control message."

This case is often seen in production which is labeled as "no socket"
because the 4-tuple of skb cannot successfully match a right and proper
socket. It is depicted by this sentecne "If the connection does not exist
(CLOSED), then a reset is sent in response to any incoming segment except
another reset." in 3.5.2. Reset Generation.

Signed-off-by: Jason Xing <kernelxing@tencent.com>