add 'Unexpected SYN in the window' reset reason
RFC 9293 3.10.7.4. Other States says:
"Fourth, check the SYN bit
...
For implementations that do not follow RFC 5961, the original behavior
described in RFC 793 follows in this paragraph. If the SYN is in the
window it is an error: send a reset"

It's worth to know that the Linux kernel doesn't implement the above.
Linux kernel chose to follow RFC 5961 so the RST packet would not
be sent in this case.

Signed-off-by: Jason Xing <kernelxing@tencent.com>